Discord Music Bot/along-bot/requirements.txt
bs4
pandas
discord.py==1.7.3
